pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "genice3_svg/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["genice3_svg"]

[project]
dynamic = ["version"]
name = "genice3-svg"
dependencies = [
    "genice3>=3.0.a0",
    "svgwrite>=1.4",
    "Pillow>=10.0",
    "numpy>=1.24",
    "cycless>=0.4.2",
    "networkx>=2.0",
]

[project.entry-points."genice3_exporter"]
svg = "genice3_svg.exporter.svg"
png = "genice3_svg.exporter.png"

[project.urls]
Homepage = "https://github.com/vitroid/genice3-svg"

[tool.genice.urls]
repository = "https://github.com/vitroid/GenIce"
logo = "https://raw.githubusercontent.com/vitroid/GenIce/develop/logo/genice-v0.png"

# poetry self add poetry-version-plugin
[tool.poetry-version-plugin]
source = "init"

[tool.poetry]
name = "genice3-svg"
version = "0.0.0"
description = "SVG and PNG visualization plugins for GenIce3."
authors = ["vitroid <vitroid@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
genice3 = "^3.0.a0"
svgwrite = "^1.4"
Pillow = "^10.0"
numpy = "^1.24"
cycless = "^0.4.2"
networkx = "^2.0"

[tool.poetry.group.dev.dependencies]
jinja2 = "^3.1.3"
toml = "^0.10.2"
jupyter = "^1.0.0"

[tool.poetry.plugins.genice3_exporter]
svg = "genice3_svg.exporter.svg"
png = "genice3_svg.exporter.png"

